use dataset types from veda-ui and move transform logic
