+++ app/(datasets)/data-catalog/page.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { getDatasets } from 'app/content/utils/mdx';
 import { Suspense } from 'react';
 import Catalog from './catalog';
-import { transformData } from '@helpers/data';
+import { transformDatasets } from '@helpers/data';
 
 export default function Page() {
   const posts: any[] = getDatasets(); // @TODO: Revist type here, should use data types from veda-ui
-  const transformed = transformData(posts);
+  const transformed = transformDatasets(posts);
 
   return (
     <section>
